+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0 (2013-10-15)
+
+- Documentation in README.md
+- Added Mocha tests
+- Added meta data to data event arguments
+- Minor bugfixes
+
 ## 0.1.0 (2013-10-10)
 
-First release!
+- First release!